==> scripts/4_removal_anomaly.ipynb <==
# Remove anomaly 

import pandas as pd
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS
from scipy.stats import pearsonr
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

Based on the previous measures, I remove those outliers: (1) articles longer than 10000 characters (aiming to exclude overly lengthy articles) (2) compression_ration that is smaller or equal to 0.01 and bigger than 0.08 (narrows the dataset to articles with summaries that are neither too brief nor excessively detailed relative to the original content).

After Removal, the correlation coefficient increased to 0.09 (Before removal: The correlation coefficient was 0.05) This indicates a small improvement in the correlation coefficient, suggesting a slightly stronger linear relationship between the lengths of the articles and their summaries after removing outliers.

For P value, it decreased to 0.11, which was 0.40 before removal. This indicates that the correlation observed is statistically more meaningful.

# Load the existing data from CSV
df = pd.read_csv('articles_with_preprocessed_and_measures.csv')

# Remove articles longer than 10000 characters
df = df[df['preprocessed_articles'].apply(len) <= 10000]

# Remove compression_ration that is smaller or equal to 0.01 and bigger than 0.08
df = df[(df['compression_ratio'] >= 0.01) & (df['compression_ratio'] < 0.08)]

# Calculate correlation and p-value once for the entire dataset
article_lengths = df['preprocessed_articles'].apply(len)
highlight_lengths = df['preprocessed_highlights'].apply(len)
correlation, p_value = pearsonr(article_lengths, highlight_lengths)

# Assign the correlation and p-value to each row in new columns
df['correlation'] = correlation
df['p_value'] = p_value

#df.drop(columns=['highlights', 'article', 'preprocessed_highlights', 'preprocessed_articles'], inplace=True)
df.drop('cosine_similarity', axis=1, inplace=True)
df.to_csv('articles_with_removal.csv', index=False)

print(f"Correlation: {correlation}, P value: {p_value}")
print("Data saved to 'articles_with_removal.csv'")
