==> scripts/6_splitdata.ipynb <==
# Split data

import pandas as pd
from sklearn.model_selection import train_test_split

df = pd.read_csv('top_200_article_summary_pairs.csv')

# Prepare the data
textes = df['articles'].tolist() 
labels = df['highlights'].tolist()

# Split the data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(
    textes, labels, test_size=0.2, random_state=0)  # Random state for reproducibility

#save the splits to CSV files
train_df = pd.DataFrame({'articles': X_train, 'highlights': y_train})
test_df = pd.DataFrame({'articles': X_test, 'highlights': y_test})

train_df.to_csv('train_article_summary_pairs.csv', index=False)
test_df.to_csv('test_article_summary_pairs.csv', index=False)

print("Training and testing data have been saved.")
